+++ src/lib/Scenes/Fair/Fair.tsx
@@ -10,6 +10,7 @@ import { FairArtworksRenderer as FairArtworksScreen } from "./Screens/FairArtwor
 import { FairBoothContainer as FairBoothScreen } from "./Screens/FairBooth"
 import { FairDetailContainer as FairDetailScreen } from "./Screens/FairDetail"
 import { FairExhibitorsRenderer as FairExhibitorsScreen } from "./Screens/FairExhibitors"
+import { FairMoreInfoContainer as FairMoreInfoScreen } from "./Screens/FairMoreInfo"
 
 import { Fair_fair } from "__generated__/Fair_fair.graphql"
 
@@ -65,6 +66,12 @@ export class Fair extends React.Component<Props> {
     this.navigate({ component: ShowArtworksScreen, props })
   }
 
+  handleViewMoreInfoPressed = () => {
+    this.navigate({
+      component: FairMoreInfoScreen,
+    })
+  }
+
   render() {
     return (
       <Theme>
@@ -82,6 +89,7 @@ export class Fair extends React.Component<Props> {
               onViewAllArtworksPressed: this.handleViewAllArtworksPressed,
               onViewAllArtistsPressed: this.handleViewAllArtistsPressed,
               onViewFairBoothPressed: this.handleViewFairBoothPressed,
+              onViewMoreInfoPressed: this.handleViewMoreInfoPressed,
             },
           }}
           style={{ flex: 1 }}